secure url

+++ app/Http/Controllers/AuthController.php
@@ -13,13 +13,14 @@ class AuthController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function auth()
-    {
-        return view('dashboard.auth');
-    }
 
-    public function index()
+    public function index(Request $request)
     {
+        $user = Auth::user();
+
+        if (!$user || !$request->user()->hasPermission('users_access')) {
+            return redirect()->route('dashboard.index')->with('error', 'Unauthorized access');
+        }
         $users = User::all();
         return view('dashboard.allusers', compact('users'));
     }
@@ -27,8 +28,13 @@ class AuthController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
+        $user = Auth::user();
+
+        if (!$user || !$request->user()->hasPermission('users_access')) {
+            return redirect()->route('dashboard.index')->with('error', 'Unauthorized access');
+        }
         $roles = Role::all(); // ✅ استرجاع كل الأدوار
         return view('dashboard.adduser', compact('roles'));
     }
@@ -39,6 +45,12 @@ class AuthController extends Controller
      */
     public function store(Request $request)
     {
+        $user = Auth::user();
+
+        if (!$user || !$request->user()->hasPermission('users_access')) {
+            return redirect()->route('dashboard.index')->with('error', 'Unauthorized access');
+        }
+
         $request->validate([
             'name' => 'required|string|max:255',
             'surname' => 'required|string|max:255',
@@ -80,8 +92,15 @@ class AuthController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(string $id, Request $request)
     {
+
+        $user = Auth::user();
+
+        if (!$user || !$request->user()->hasPermission('users_access')) {
+            return redirect()->route('dashboard.index')->with('error', 'Unauthorized access');
+        }
+
         return view('dashboard.auth.show', [
             'user' => User::findOrFail($id)
         ]);
@@ -90,8 +109,14 @@ class AuthController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(string $id, Request $request)
     {
+        $user = Auth::user();
+
+        if (!$user || !$request->user()->hasPermission('users_access')) {
+            return redirect()->route('dashboard.index')->with('error', 'Unauthorized access');
+        }
+
         $user = User::findOrFail($id);
         return view('dashboard.auth.edit', compact('user'));
     }
@@ -101,6 +126,11 @@ class AuthController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $user = Auth::user();
+
+        if (!$user || !$request->user()->hasPermission('users_access')) {
+            return redirect()->route('dashboard.index')->with('error', 'Unauthorized access');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
             'surname' => 'required|string|max:255',
@@ -140,14 +170,25 @@ class AuthController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $id, Request $request)
     {
+        $user = Auth::user();
+
+        if (!$user || !$request->user()->hasPermission('users_access')) {
+            return redirect()->route('dashboard.index')->with('error', 'Unauthorized access');
+        }
+
         $user = User::findOrFail($id);
         $user->delete();
 
-        return response()->json([
-            'message' => 'User deleted successfully'
-        ], 200);
+        return redirect()->route('dashboard.roles.index')
+            ->with('success', 'تم حذف المستعمل بنجاح.');
+    }
+
+
+    public function auth()
+    {
+        return view('dashboard.auth');
     }
 
     /**
